server/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.28.2
werkzeug==2.2.3
gunicorn==21.2.0
groq==0.4.1
